+++ src/components/modules/blockManager.ts
@@ -337,19 +337,26 @@ export default class BlockManager extends Module {
    * Update Block data.
    *
    * Currently we don't have an 'update' method in the Tools API, so we just create a new block with the same id and type
-   * Should not trigger 'block-removed' or 'block-added' events
+   * Should not trigger 'block-removed' or 'block-added' events.
+   *
+   * If neither data nor tunes is provided, return the provided block instead.
    *
    * @param block - block to update
-   * @param data - new data
+   * @param data - (optional) new data
+   * @param tunes - (optional) tune data
    */
-  public async update(block: Block, data: Partial<BlockToolData>): Promise<Block> {
+  public async update(block: Block, data?: Partial<BlockToolData>, tunes?: {[name: string]: BlockTuneData}): Promise<Block> {
+    if (!data && !tunes) {
+      return block;
+    }
+
     const existingData = await block.data;
 
     const newBlock = this.composeBlock({
       id: block.id,
       tool: block.name,
-      data: Object.assign({}, existingData, data),
-      tunes: block.tunes,
+      data: Object.assign({}, existingData, data ?? {}),
+      tunes: tunes ?? block.tunes,
     });
 
     const blockIndex = this.getBlockIndex(block);
@@ -544,7 +551,7 @@ export default class BlockManager extends Module {
        * If first Block was removed, insert new Initial Block and set focus on it`s first input
        */
       if (!this.blocks.length) {
-        this.currentBlockIndex = -1;
+        this.unsetCurrentBlock();
 
         if (addLastBlock) {
           this.insert();
@@ -591,7 +598,7 @@ export default class BlockManager extends Module {
       this._blocks.remove(index);
     }
 
-    this.currentBlockIndex = -1;
+    this.unsetCurrentBlock();
     this.insert();
     this.currentBlock.firstInput.focus();
   }
@@ -873,7 +880,7 @@ export default class BlockManager extends Module {
    * Sets current Block Index -1 which means unknown
    * and clear highlights
    */
-  public dropPointer(): void {
+  public unsetCurrentBlock(): void {
     this.currentBlockIndex = -1;
   }
 
@@ -895,7 +902,7 @@ export default class BlockManager extends Module {
 
     await queue.completed;
 
-    this.dropPointer();
+    this.unsetCurrentBlock();
 
     if (needToAddDefaultBlock) {
       this.insert();