+++ src/components/utils/popover/components/hint/hint.const.ts
@@ -10,7 +10,8 @@ const className = bem('ce-hint');
  */
 export const css = {
   root: className(),
-  alignedLeft: className(null, 'align-left'),
+  alignedStart: className(null, 'align-left'),
+  alignedCenter: className(null, 'align-center'),
   title: className('title'),
   description: className('description'),
 };