+++ test/cypress/tests/modules/InlineToolbar.cy.ts
@@ -19,7 +19,7 @@ describe('Inline Toolbar', () => {
       .find('.ce-paragraph')
       .selectText('block');
 
-    cy.get('[data-cy="inline-toolbar"]')
+    cy.get('[data-cy="inline-toolbar"] .ce-popover__container')
       .should('be.visible')
       .then(($toolbar) => {
         const editorWindow = $toolbar.get(0).ownerDocument.defaultView;
@@ -28,7 +28,7 @@ describe('Inline Toolbar', () => {
         const range = selection.getRangeAt(0);
         const rect = range.getBoundingClientRect();
 
-        expect($toolbar.offset().left).to.closeTo(rect.left, 1);
+        expect($toolbar.offset().left).to.be.closeTo(rect.left, 1);
       });
   });
 
@@ -61,7 +61,7 @@ describe('Inline Toolbar', () => {
           .selectTextByOffset([firstLineWrapIndex - 5, firstLineWrapIndex - 1]);
       });
 
-    cy.get('[data-cy="inline-toolbar"]')
+    cy.get('[data-cy="inline-toolbar"] .ce-popover__container')
       .should('be.visible')
       .then(($toolbar) => {
         cy.get('@blockWrapper')
@@ -71,9 +71,45 @@ describe('Inline Toolbar', () => {
             /**
              * Toolbar should be aligned with right side of text column
              */
-            expect($toolbar.offset().left + $toolbar.width()).to.closeTo(blockWrapperRect.right, 3);
+            expect($toolbar.offset().left + $toolbar.width()).to.closeTo(blockWrapperRect.right, 10);
           });
       });
+  });
+
+  it('should not submit form nesting editor when inline tool clicked', () => {
+    cy.createEditor({
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Some text',
+            },
+          },
+        ],
+      },
+    });
+
+    const onSubmit = cy.stub();
+
+    cy.document().then(doc => {
+      const form = doc.createElement('form');
+
+      form.onsubmit = onSubmit;
+      doc.body.appendChild(form);
+
+      /* Move editor to form */
+      form.appendChild(doc.getElementById('editorjs'));
+
+      cy.get('[data-cy=editorjs]')
+        .find('.ce-paragraph')
+        .selectText('Some text');
+
+      cy.get('[data-item-name=bold]')
+        .click();
+
+      expect(onSubmit).to.be.not.called;
+    });
   });
 
   describe('Conversion toolbar', () => {
@@ -100,11 +136,12 @@ describe('Inline Toolbar', () => {
         .find('.ce-paragraph')
         .selectText('Some text');
 
-      cy.get('[data-cy=conversion-toggler]')
+      cy.get('[data-item-name=convert-to]')
         .click();
 
       cy.get('[data-cy=editorjs]')
-        .find('.ce-conversion-tool[data-tool=header]')
+        .get('.ce-inline-toolbar')
+        .find('.ce-popover-item[data-item-name=header]')
         .click();
 
       cy.get('[data-cy=editorjs]')