+++ types/api/blocks.d.ts
@@ -1,6 +1,8 @@
+import Block from '../../src/components/block';
 import {OutputBlockData, OutputData} from '../data-formats/output-data';
 import {BlockToolData, ToolConfig} from '../tools';
 import {BlockAPI} from './block';
+import {BlockTuneData} from '../block-tunes/block-tune-data';
 
 /**
  * Describes methods to manipulate with Editor`s blocks
@@ -72,6 +74,13 @@ export interface Blocks {
   getBlockIndex(blockId: string): number;
 
   /**
+   * Get Block API object by html element
+   *
+   * @param element - html element to get Block by
+   */
+  getBlockByElement(element: HTMLElement): BlockAPI | undefined;
+
+  /**
    * Mark Block as stretched
    * @param {number} index - Block to mark
    * @param {boolean} status - stretch status
@@ -134,9 +143,10 @@ export interface Blocks {
    * Updates block data by id
    *
    * @param id - id of the block to update
-   * @param data - the new data. Can be partial.
+   * @param data - (optional) the new data. Can be partial.
+   * @param tunes - (optional) tune data
    */
-  update(id: string, data: Partial<BlockToolData>): Promise<BlockAPI>;
+  update(id: string, data?: Partial<BlockToolData>, tunes?: {[name: string]: BlockTuneData}): Promise<BlockAPI>;
 
   /**
    * Converts block to another type. Both blocks should provide the conversionConfig.